Model verification - User defined queries
A user may now specify custom queries that will be verified during verification phase. These queries are also checked if they use correct notation, location and edges using the correct naming and so on.

This feature is highly relevant if a user specifies a more accurate model as a representation of the real world.

Tested the custom query readings to a small degree.

Also fancy console colors for printing Success, Warning and Failure in different colors.