Liveness and Bound queries + small naive world fix
Added query for liveness of the controller as it is the main template and also added queries for Bound templates.

C arrays are 0 indexed therefore the select statement should be pinCount - 1 so it would look something like this with a model with 1 pin (chan outPin[1][2]):
i : int[0, 0] for outPin[i][1] as i : int[0, 1] would be out of range.